minor fixes - location change of helper function
